// src/app/models/user.model.ts
export interface User {
  id: any;
  role_type: string;
  title: string;
  first_name: string;
  last_name: string;
  email: string;
  email_signature: string;
  password: string;
  password_reset: string;
  mobile: string;
  image: string;
  job_title: string;
  active: string;
  status_id: number;
  is_admin: boolean;
  role_id: number;
  username: string;
  //     "id": 0,
  //   "role_type": 0,
  //   "title": "string",
  //   "first_name": "string",
  //   "last_name": "string",
  //   "email": "string",
  //   "email_signature": "string",
  //   "password": "string",
  //   "password_reset": true,
  //   "mobile": "string",
  //   "image": "string",
  //   "job_title": "string",
  //   "active": 0,
  //   "status_id": 0,
  //   "is_admin": true
}